extract common db setup
